Skating Board Project
Project created to display how hockey teams are assigned locker rooms.
The board saves data and needs to be reupdated weekly to stay up todate.

The Main file is the board and Client file is used to adjust
information.

The Client file doesn't need to be on the same computer as the board.
Data is passed over when the client connects to the server. After the
data has been adjusted, the user sends the data back to the board which
updates the information being displayed.

The next event is highlighted in blue to help attract attention. An
emergency message and a promotional message can be displayed.

Users can add pictures to a Pictures folder and the program will display
all pictures in the folder (requires the program to be rebooted if new
pictures are added).

The Board/Server saves the data when closed and reupdates the data when
open. The User can save data and reused if the event occurs more then
once.