Fixed bug with deleting on next day.
There was a bug in the server, the data sent to the board was saved
within the server file. If a user would send data to the board, the data
would be saved in the Server file. If the board would change the next
day, the data for the board would be delete the last day. Although, it
wouldn't update the data for the Server. Therefore, when the board would
send back the data to a client, the data previously sent would be sent
back without the manipulated data from the board.