Removed Pictures and the Error Log
The Error Log didn't work, it would only collect the first Error and
it's an inefficient way to catch errors during production.

The Pictures didn't work very well, they would disappear over time
without a way to tell why. The pictures would randomly disappear without
giving an error. Since this was unnecessary and added in to learn and
look cool. It was removed because it was causing more issues then what
it's worth.